Task 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,8 +10,8 @@
 int main(void)
 {
 	int a = 0;
-	int b = 1;
-	int c = 2;
+	long b = 1;
+	long c = 2;
 
 	while (a < 50)
 	{
